requirements.txt
pydantic>=2.6
pydantic-settings>=2.2
PyYAML>=6.0
websockets>=12.0
websocket-client>=1.8
pandas>=2.2
numpy>=1.26
matplotlib>=3.9
pytest>=8.2
pytest-asyncio>=0.23
ruff>=0.5
typing-extensions>=4.12
